feat(TagOverflow): implement useOverflow hook
